Create simple input and output
stimated time
5-10 minutes

Level of difficulty
Easy

Objectives
becoming familiar with the inputting and outputting of data in Python;
evaluating simple expressions.
Scenario
Your task is to complete the code in order to evaluate the results of four basic arithmetic operations.

The results have to be printed to the console.

You may not be able to protect the code from a user who wants to divide by zero. That's okay, don't worry about it for now.

Test your code - does it produce the results you expect?

We won't show you any test data - that would be too simple.