Create FUnctions -seeing if year is leap year
Estimated time
10-15 minutes

Level of difficulty
Easy

Objectives
Familiarize the student with:

projecting and writing parameterized functions;
utilizing the return statement;
testing the functions.
Scenario
Your task is to write and test a function which takes one argument (a year) and returns True if the year is a leap year, or False otherwise.

The seed of the function is already shown in the skeleton code in the editor.

Note: we've also prepared a short testing code, which you can use to test your function.

The code uses two lists ‒ one with the test data, and the other containing the expected results. The code will tell you if any of your results are invalid.