Create Classes first lab- Counting stack
Level of difficulty
Easy/Medium

Objectives
improve the student's skills in defining classes;
using existing classes to create new classes equipped with new functionalities.
Scenario
We've showed you recently how to extend Stack possibilities by defining a new class (i.e., a subclass) which retains all inherited traits and adds some new ones.

Your task is to extend the Stack class behavior in such a way so that the class is able to count all the elements that are pushed and popped (we assume that counting pops is enough). Use the Stack class we've provided in the editor.

Follow the hints:

introduce a property designed to count pop operations and name it in a way which guarantees hiding it;
initialize it to zero inside the constructor;
provide a method which returns the value currently assigned to the counter (name it get_counter()).
Complete the code in the editor. Run it to check whether your code outputs 100.