Create is list empty
5-30 minutes

Level of difficulty
Easy/Medium

Objectives
improving the student's skills in defining subclasses;
adding a new functionality to an existing class.
Scenario
Your task is to slightly extend the Queue class' capabilities. We want it to have a parameterless method that returns True if the queue is empty and False otherwise.

Complete the code we've provided in the editor. Run it to check whether it outputs a similar result to ours.

Below you can copy the code we used in the previous lab:

Check

Expected output
1
dog
False
Queue empty